+++ app/src/app/components/contents/manageAccess/AddRoleButton.tsx
@@ -1,27 +1,9 @@
-import { zodResolver } from "@hookform/resolvers/zod";
 import { SquarePlus } from "lucide-react";
 import { useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { GLOBAL_PERMISSIONS } from "../../../../lib/constants";
-import { Button } from "../../Base/Button";
-import { Checkbox } from "../../Base/Checkbox";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger
-} from "../../Base/Dialog";
-import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "../../Base/Form";
-import { Input } from "../../Base/Input";
-import { fetcher } from "../../../../lib/fetch";
-import appConfig from "../../../../config/appConfig";
-import { ApiError } from "../../../../lib/errors";
-import { HttpStatus } from "../../../../lib/types/HttpStatus";
 import { KeyedMutator } from "swr";
+import { Button } from "../../Base/Button";
+import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from "../../Base/Dialog";
+import { AddRoleForm } from "./AddRoleForm";
 import { RoleWithRelationships } from "./types/RoleWithRelationships";
 
 interface AddRoleButtonProps {
@@ -29,37 +11,7 @@ interface AddRoleButtonProps {
 }
 export const AddRoleButton = ({ mutate }: AddRoleButtonProps) => {
   const [open, setOpen] = useState(false);
-  const [fetchError, setFetchError] = useState("");
 
-  const formSchema = z.object({
-    name: z.string().min(1),
-    permissionNames: z.array(z.string())
-  });
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      permissionNames: []
-    }
-  });
-  const onSubmit = async (values: z.infer<typeof formSchema>) => {
-    try {
-      await fetcher({
-        url: `${appConfig.apiUrl()}/role`,
-        body: values,
-        method: "POST"
-      });
-      form.reset();
-      setOpen(false);
-      mutate();
-    } catch (error) {
-      console.error(error);
-      if (error instanceof ApiError && error.status === HttpStatus.BadRequest) {
-        return form.setError("name", { message: error.message });
-      }
-      setFetchError("An unexpected error occurred. Please try again.");
-    }
-  };
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger asChild>
@@ -72,71 +24,7 @@ export const AddRoleButton = ({ mutate }: AddRoleButtonProps) => {
         <DialogHeader>
           <DialogTitle>Create new Role</DialogTitle>
         </DialogHeader>
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-3">
-            <FormField
-              control={form.control}
-              name="name"
-              render={({ field }) => (
-                <FormItem>
-                  <FormLabel>Name</FormLabel>
-                  <FormControl>
-                    <Input autoComplete="name" placeholder="enter role name..." {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name="permissionNames"
-              render={() => (
-                <FormItem>
-                  <div className="mb-3">
-                    <FormLabel>Permissions</FormLabel>
-                    <FormDescription className="text-xs">
-                      Select the permissions that will be assigned to this role.
-                    </FormDescription>
-                  </div>
-                  {GLOBAL_PERMISSIONS.map((permission, idx) => (
-                    <FormField
-                      key={`permission-${idx}`}
-                      control={form.control}
-                      name="permissionNames"
-                      render={({ field }) => {
-                        return (
-                          <FormItem key={`permission-${idx}`} className="flex flex-row items-start space-x-3 space-y-0">
-                            <FormControl>
-                              <Checkbox
-                                checked={field.value?.includes(permission)}
-                                onCheckedChange={(checked) => {
-                                  return checked
-                                    ? field.onChange([...field.value, permission])
-                                    : field.onChange(field.value?.filter((value) => value !== permission));
-                                }}
-                              />
-                            </FormControl>
-                            <FormLabel className="font-normal">{permission}</FormLabel>
-                          </FormItem>
-                        );
-                      }}
-                    />
-                  ))}
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <DialogFooter className="sm:justify-end gap-1">
-              {fetchError && <div className="text-xs text-red-500">{fetchError}</div>}
-              <DialogClose asChild>
-                <Button type="button" variant="secondary" onClick={() => form.reset()}>
-                  Cancel
-                </Button>
-              </DialogClose>
-              <Button type="submit">Create</Button>
-            </DialogFooter>
-          </form>
-        </Form>
+        <AddRoleForm mutate={mutate} setOpen={setOpen} />
       </DialogContent>
     </Dialog>
   );
